Moved related functions into same file
