+++ admin/editPortfolio.php
@@ -2,7 +2,7 @@
 require ('dbStream.php');
 $stmtPortfolio = $db->query('SELECT `id`,`title`,`link`,`github`,`image`,`description` FROM `portfolio`;');
 $dataPortfolio = $stmtPortfolio->fetchAll();
-require('dbDisplay.php');
+require('../functions/db.php');
 ?>
 <!DOCTYPE html>
 <html lang="en">
