+++ contact.html
@@ -13,7 +13,7 @@
         <a href="." class="homeButton">James</a>
         <nav class="topNav">
             <a href="about.html">About</a>
-            <a href="portfolio.html">Portfolio</a>
+            <a href="portfolio.php">Portfolio</a>
             <a href="contact.html">Contact</a>
         </nav>
         <nav class="socialLinks">
